refactor(devframe)!: slim devframe/node to the context surface; drop isObject
Per API-surface review:

- Remove `isObject` from the public API entirely (unused internally; trivial to
  inline).
- Move the server-assembly primitives `createH3DevframeHost` (+ its options) and
  `startHttpAndWs` (+ `StartedServer`, options) from `devframe/node` to the
  unstable `devframe/internal` surface. `devframe/node` now exposes only the
  context-building API: `createHostContext`, `createStorage`, `RpcFunctionsHost`.
- `createDevServer` stays — it's the standalone-server adapter (built on
  `initDevframe`, used by `createCac` and consumer CLIs), a different layer from
  the `initDevframe` handler core.

The hub, plugin test harnesses, the a11y demo, storybook, and host examples now
import `createH3DevframeHost` / `startHttpAndWs` from `devframe/internal`.
Migration guide, plan, and tsnapi snapshots updated.
