+++ examples/files-inspector/tests/_utils.ts
@@ -1,4 +1,4 @@
-import type { StartedServer } from 'devframe/node'
+import type { StartedServer } from 'devframe/internal'
 import { existsSync } from 'node:fs'
 import { mkdtemp, writeFile } from 'node:fs/promises'
 import os from 'node:os'
@@ -7,11 +7,8 @@ import { fileURLToPath } from 'node:url'
 import {
   DEVFRAME_CONNECTION_META_FILENAME,
 } from 'devframe/constants'
-import {
-  createH3DevframeHost,
-  createHostContext,
-  startHttpAndWs,
-} from 'devframe/node'
+import { createH3DevframeHost, startHttpAndWs } from 'devframe/internal'
+import { createHostContext } from 'devframe/node'
 import { mountStaticHandler } from 'devframe/utils/serve-static'
 import { getPort } from 'get-port-please'
 import { H3 } from 'h3'
