+++ packages/devframe/src/node/index.ts
@@ -1,12 +1,13 @@
-// Node-side public API for hosts that wire up their own runtime — the
-// server-assembly surface (`createHostContext` → `createH3DevframeHost` →
-// `startHttpAndWs`) and storage.
+// Node-side public API for building a devframe context: `createHostContext`
+// (the context assembler) and `createStorage`.
 //
-// Everything lower-level lives at `devframe/internal` (an explicitly-unstable
+// The server-assembly primitives (`createH3DevframeHost`, `startHttpAndWs`) and
+// everything lower-level live at `devframe/internal` (an explicitly-unstable
 // cross-package surface): the transport-agnostic RPC core, the agent host, the
-// instance registry (host self-registration + live discovery), and the
-// `isObject` / `normalizeHttpServerUrl` helpers. The diagnostics/services/views
-// hosts, the streaming/shared-state/scope/settings factories, and the internal
+// instance registry, and the `normalizeHttpServerUrl` helper. Application code
+// serves a devframe through the adapters (`createDevServer`, `createBuild`,
+// `createCac`) or `devframe/initiate`. The diagnostics/services/views hosts,
+// the streaming/shared-state/scope/settings factories, and the internal
 // host-URL helpers stay fully internal (relative imports only).
 // `toAgentToolName` lives at `devframe/utils/agent-tool-name`.
 export * from './context'
@@ -14,6 +15,4 @@ export * from './context'
 // `RpcFunctionsHost` type so consumers can type/cast `ctx.rpc` without
 // pulling in the implementation's `@internal` members.
 export type { RpcFunctionsHost } from './host-functions'
-export * from './host-h3'
-export * from './server'
 export * from './storage'
