+++ packages/hub/src/node/__tests__/context.test.ts
@@ -2,7 +2,8 @@ import type { DevframeDockEntry } from '../../types/docks'
 import { mkdtempSync } from 'node:fs'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
-import { createHostContext, startHttpAndWs } from 'devframe/node'
+import { startHttpAndWs } from 'devframe/internal'
+import { createHostContext } from 'devframe/node'
 import { getInternalContext } from 'devframe/node/hub-internals'
 import { describe, expect, it, vi } from 'vitest'
 import { createHubContext } from '../context'
