+++ packages/hub/src/node/initiate.ts
@@ -1,4 +1,4 @@
-import type { StartedServer } from 'devframe/node'
+import type { StartedServer } from 'devframe/internal'
 import type { DevframeAuthHandler } from 'devframe/node/auth'
 import type { BunWsTier } from 'devframe/rpc/transports/ws-bun'
 import type { WsOriginRegistry } from 'devframe/rpc/transports/ws-server'
@@ -10,7 +10,7 @@ import type { MountDevframeOptions } from './mount-devframe'
 import { readFile } from 'node:fs/promises'
 import process from 'node:process'
 import { DEVFRAME_CONNECTION_META_FILENAME, DEVFRAME_DOCK_IMPORTS_FILENAME, DEVFRAME_MCP_ROUTE, DEVFRAME_WS_ROUTE } from 'devframe/constants'
-import { createH3DevframeHost, startHttpAndWs } from 'devframe/node'
+import { createH3DevframeHost, startHttpAndWs } from 'devframe/internal'
 import { createInteractiveAuth } from 'devframe/recipes/interactive-auth'
 import { mountStaticHandler } from 'devframe/utils/serve-static'
 import { getPort } from 'get-port-please'
