+++ plans/032-0.9-public-api-surface-reduction.md
@@ -95,18 +95,21 @@ Net effect (after maintainer follow-ups):
 - Dead-subpath removals are `utils/promise` and `utils/scope` only. `ws-bun` is
   kept (hub uses it); `utils/hash` is kept (maintainer request); `embedded` is
   kept (maintainer request — a named, discoverable adapter).
-- The public `devframe/node` barrel is reduced to the server-assembly surface
-  only: `createHostContext`, `createH3DevframeHost`, `startHttpAndWs`
-  (+ `StartedServer`), `createStorage`, and the `RpcFunctionsHost` type.
+- The public `devframe/node` barrel is reduced to the context surface only:
+  `createHostContext` (+ options), `createStorage` (+ options), and the
+  `RpcFunctionsHost` type. Serving is done through the adapters or
+  `devframe/initiate`.
 - A new **`devframe/internal`** entry (an explicitly-unstable cross-package
   surface) holds the low-level primitives first-party integrations need:
+  `createH3DevframeHost`, `startHttpAndWs` (+ `StartedServer`),
   `createContextRpcServer`, `DevframeAgentHost`, `coerceAgentPositionalArgs`,
   the instance registry (`registerDevframeInstance` / `listLiveDevframeInstances`
-  + types), and `isObject` / `normalizeHttpServerUrl`. The hub, the inspect
-  plugin, and the host examples import from there. The remaining node internals
-  (diagnostics/services/views hosts, the streaming/shared-state/scope/settings
-  factories, `toDialableHost`/`formatHostForUrl`) stay fully internal (relative
-  imports only).
+  + types), and `normalizeHttpServerUrl`. The hub, the inspect plugin, the host
+  examples, and every plugin's test harness import from there.
+- `isObject` is **removed entirely** (unused internally; trivial to inline).
+- The remaining node internals (diagnostics/services/views hosts, the
+  streaming/shared-state/scope/settings factories, `toDialableHost`/
+  `formatHostForUrl`) stay fully internal (relative imports only).
 
 ### ⚠️ One item to confirm during review
