+++ plugins/a11y/demo/server.mjs
@@ -23,7 +23,8 @@ import { createServer } from 'node:http'
 import process from 'node:process'
 import { fileURLToPath } from 'node:url'
 import { DEVFRAME_CONNECTION_META_FILENAME } from 'devframe/constants'
-import { createH3DevframeHost, createHostContext, startHttpAndWs } from 'devframe/node'
+import { createH3DevframeHost, startHttpAndWs } from 'devframe/internal'
+import { createHostContext } from 'devframe/node'
 import { mountStaticHandler } from 'devframe/utils/serve-static'
 import { getPort } from 'get-port-please'
 import { H3, toNodeHandler } from 'h3'
