+++ plugins/assets/test/_utils.ts
@@ -1,12 +1,13 @@
 import type { DevframeNodeContext } from 'devframe'
-import type { StartedServer } from 'devframe/node'
+import type { StartedServer } from 'devframe/internal'
 import type { AssetsDevframeOptions } from '../src/index'
 import { mkdtemp, rm } from 'node:fs/promises'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import process from 'node:process'
 import { createRpcStreamingClientHost } from 'devframe/client'
-import { createH3DevframeHost, createHostContext, startHttpAndWs } from 'devframe/node'
+import { createH3DevframeHost, startHttpAndWs } from 'devframe/internal'
+import { createHostContext } from 'devframe/node'
 import { createRpcClient } from 'devframe/rpc/client'
 import { createWsRpcChannel } from 'devframe/rpc/transports/ws-client'
 import { mountStaticHandler } from 'devframe/utils/serve-static'
