+++ plugins/data-inspector/src/inject/index.ts
@@ -42,7 +42,8 @@ import { mkdirSync, rmSync, writeFileSync } from 'node:fs'
 import { homedir } from 'node:os'
 import { dirname, join } from 'node:path'
 import process from 'node:process'
-import { createHostContext, startHttpAndWs } from 'devframe/node'
+import { startHttpAndWs } from 'devframe/internal'
+import { createHostContext } from 'devframe/node'
 import { createInteractiveAuth } from 'devframe/recipes/interactive-auth'
 import { randomToken } from 'devframe/utils/crypto-token'
 import { getPort } from 'get-port-please'
