+++ storybook/src/hub.ts
@@ -11,7 +11,7 @@ import { fileURLToPath } from 'node:url'
 import { createHubContext, mountDevframe } from '@devframes/hub/node'
 import terminalsDevframe from '@devframes/plugin-terminals'
 import { DEVFRAME_CONNECTION_META_FILENAME } from 'devframe/constants'
-import { startHttpAndWs } from 'devframe/node'
+import { startHttpAndWs } from 'devframe/internal'
 import { serveStaticNodeMiddleware } from 'devframe/utils/serve-static'
 import { getPort } from 'get-port-please'
 import { dirname, join } from 'pathe'
